Teste/task_manager/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
